Headlandturn adjustments
Before, big headers turned to much outwards the field after the turn. Now we stay a bit more inside, while the smaler headers still do a clean work.
Tested with the smalest Maize cutter and the biggest corn cutter.
Maybe have a look to get it working depending on the workwidth, to have less problems with bigger and smaler headers difference.