Push Release 7.0.0.18
- Some changes in the Job starting and stopping
- Show course start/end on plot #191
- Synchronize the cp job parameter between the in game menu and the mini gui
- Turn fix and improvement for #461
- Added Interface function for AutoDrive
- Added the Bale wrapper *1
- Added Bale Loaders *2
- Error popup on course generator error #472
- Added a field margin setting to generator screen *3
- Articulated vehicle reversing fix #230 #329 #424
- Also fixes unfold for Ropa Harvester
- More turn fixes and improovements
- Attached harvester fix #370
- Smal translation changes

Due to the confusion about our 'mini HUD', changed the actual display keybind to a proper name. At this point a huge thank you to our translators for doing such a great job.

*1
With the 'start at' option set to 'bale collect/wrap', and with a bale wrapper, CP will find all bales on the field and
wrap them.

*2
For now, they just stop when full.

*3
This generate a new field boundary (smaller or bigger than the original) before starting the actual course generation.
This way, you can overlap the field border or get more distance to avoid obstacles that are to close to the field border.