Finding bales on custom fields
For now works only when the job is started from the AI screen as it gets
the selected custom field from the job.

When starting from the vehicle or HUD, the job does not seem to have
the selected custom field anymore.

Not sure how to best address this. If we are not on a field, we could
check for custom fields, but that won't work with overlapping fields.

#532