Last row width
If there are headlands, make the last row overlap with the
headland, and not as before, overlapping the second last row.

This hopefully will prevent collisions in this case when there
are multiple vehicles on the course, there's little chance that
one is working on the last row while another on the headland.