Update translation_ru.xml
Hello. I'm still alive. After three months of the war, I had the Internet in a new place. I don't know-how for a long time. Therefore, so far made part of the translation. And then that I will have time to send.
When are you planning to make a release? Maybe I'll try to hurry.