ModHub Version 7.1.0.0
ModHub Release 7.1.0.0

- turn off stopAtEnd setting for now
This settings is not used anywhere.
Should consider reactivate it, but it needs to be tested with combines and AutoDrive take over. As this will be a bit more complex, will do that after ModHub Release.

- fix: no convoy stop/slow down when self unloading #1690

- some small tweaks

This Version will get sent into ModHub. Keep in mind that we might get it back and have to change some small things, wich won't get a new Release Version, but the .zip file will be updated. As soon as Courseplay is on ModHub, we will continue with normal Updates.