Unloader fixes
- fixed finding the start waypoint on the combine
follow course when unloading a moving combine, made
sure the offset is set before we start looking for
the waypoint in front of us.
- unloader reconfirms the rendezvous periodically so
when the combine is waiting for it before turning into a
row where the pipe would be in the fruit, it won't give
up waiting too early.
- pipe in the fruit map fixed, the turn start waypoint at
the end of the row wasn't marked as in fruit like the
rest of the row.