Push Release 7.1.1.0
- fix for #1925
- fix for #1933 and #1931
- fix for 1914
- Changed behaviour for Cotton Harvester, fixes #1779
- #1916 should be fixed
- #1816 should be fixed
- Added some key inputs
- Introduce the 'old' new Combine unloader Mode.

The new combine unloader mode is only for combines and harvesters, choppers are not supported and probably won't ever be! Combines must be driven by Courseplay for the unloaders to work.

To unload combines, you need a tractor with an auger wagon or a trailer. With an auger wagon, Courseplay will attempt to find a trailer around the field and when the auger wagon is full, drive to that trailer and empty the auger wagon into that trailer, similar to the combine self unload.

With a trailer, Courseplay will stop when full and you have to set up AutoDrive to bring the tractor to a spot where it can empty the trailer. Alternatively, click that little steering wheel in the HUD and toggle through a list of targets to let the Giants helper handle the transport for you. Note that we did not improve the Giants helper...

Before you start the unloader, you have to select the field it will look for combines to unload. You can do this by clicking the target icon in the HUD and then selecting a Field Position.

You have to start the unloader either on the field or within 20 meters from the field edge. If the unloader is not around the field, you can also select a Target Position near the field so it'll drive to the that position before it starts working.

Combines call the unloader when their fill level reaches 80% (this is currently not configurable, but should work out just fine in most cases). You can set the 'Empty when over %. in the HUD to tell the unloader that over this fill level, it should not wait for a combine to call but go and unload the trailers immediately.