Turn fix for vehicles with reverser node
Fixed AIUtil.getReverserNode() and moveCourseBack() for the case when
there is no room on the field to make the turn without backing up
first.

Some vehicles have a reverser node which we use as a reference when
driving backwards, the reverse portion of the turn course must begin behind
this node in that case.

#1549