Don't use root vehicle to calculate the pipe offset
If the tractor is not lined up with a towed harvester or an
auger wagon, using the root vehicle (that is, the tractor) to
calculate the pipe offset results in incorrect offsets, depending
on the angle between the tractor and the implement.

Use the implement with the pipe to find the offset.