Reverse driving fix
If the first implement we checked on the back of
the tractor (root node behind tractor's node) had
nothing attached, we stopped looking and did not
find other implements.

For instance, the JD 7R had a Laforge weight attached
underneath, we ignored all towed implements, which
messed up the turn calculations.

#828 #1306 #2054