Push Release 7.1.2.0
- fix for #2094
- fix for #2100
- fix for #2092
- added a minimum offset for auger wagon to trailer of 3.8m #2092 might help with #2090
- Slightly adjusted automatic pipe movement. Should help with #2036
- Proximity sensor will now be turned off on the same side as the pipe and not only on the left side.
- Course rename enhancement #2075
- added the Bunker silo driver*
- added remaining time to the HUD for fieldwork**

*INFO:
The bunker silo mode is available if no valid fieldwork implement or trailer is attached.
The mode is preselected when a shield or a spreader is attached, with a roller you have to cycle through the starting point setting in the HUD until the bunker silo mode appears.

With a shield or spreader attached, Courseplay sets the direction to drive into the silo automatically. If the vehicle has no shield or spreader attached, you have to set the direction manually.

You can also enable/disable the waiting at the park position.
If waiting at the park position is enabled, and a vehicle with a full trailer approaches (driven by the player or by AutoDrive), the driver will automatically drive to the park position and wait until the unloader finished unloading and left the detection range near or in the bunker silo.

**INFO:
The remaining time is just a leftover time prediction for fieldwork about some fixed parameters and will like have some inaccuracy.
Refilling, standing around (e.g. harvester waiting for an unloader) is of course not part of the calculation.