fix: turns on headlands when islands present
Wide turns where the vehicle drives on one of the
headlands to the next row failed when the headland
bypassed islands elsewhere because the headland
pass number property of some waypoints were lost
when smoothing the transition to the bypass
circle.

Fix smoothing adjust the waypoint coordinates only
instead of creating a new waypoint.

#2163