Might help with issue #2295
Tries to hold onto an old reference ix from the original fieldwork waypoint, which should roughly be near the new offset waypoint for a multi tool course.
This should prevent drifting of the field work course display, as the length for a multi tool course might be different of the original course used to display the progress.