scary fix: reversing before recovery turn
When generating a straight course (forward of
reverse) which is shorter than the step given for
Course.createFromNode(), we generated a single waypoint.

Also, when the course was longer than the step, we
still generated the first waypoint further than the 'from'
distance.

Fix both cases, now there is always a waypoint at 'from' and
'to' distances from the node, and the rest is evenly distributed.