fix: multitool unloader call
In multitool mode, do not arrange a rendezvous with
an unloader at the end of a row just because it appears
that the unloader would be full in the next row.

With multitool, we have no reliable information if there
would be fruit at a given waypoint in the future.

#2373