feat: make room for a second unloader
If an unloader in idle state blocking another unloader,
we'll ask it to move as we are busy and it has
nothing to do anyway.

Such a situation can arise when the first unloader just
finished overloading to a waiting trailer and pulled
ahead a bit, waiting for a combine to call, when a second
unloader arrives to the trailer to overload, but can't
get close enough because it is blocked by the first, idle one.

#2568