fix: bale wrapper unfold
for some reason, balers/bale wrappers that don't
really unfold still have a spec_foldable, although there
is no foldable defined in the XML file. If we unfold these,
they are stuck in the unfold state, also, the check for
foldAnimTime == 1 in canContinueWork() won't work. These
all have a default 0.0001 anim time, so check if the anim
time is a reasonable value and only unfold then.

See also #2558, which broke the these balers in normal
course mode.

#2598