fix: callstack in turn radius calculation
We used getChildVehicles() when finding the turn radius, which
returns the own vehicle, not just the implements. This did not
cause any issues before, as our localToLocal returned 0 since
we were calculating it for the own vehicle, so both nodes
were the same root node.

Now with the AI direction node, it is different than the root
node and we thought it is an implement on the back.

Ignore own vehicle when checking for towed implements.