Push Release 7.2.1.7
- fix for #2563
- fix for #2564
- fix for #2616
- fix for #2519
- Fixes a bug, where the startup message was not displayed.
- Fixes a bug for proximity sensor when use reversed cabin driving.
- Added global setting for fruit destruction of helpers. (off, CP only, all Helper).
- Added a proper offset setting to the HUD for bale collect/wrap mode.
- Added new feature, where the combine automatically can pickup the header from an attached trailer:

A new setting must be activated in the vehicle settings page under combine settings for this to work.
There are two use cases:
Player startes the harvester near a field with the cutter attached on the back, either on a trailer or an integrated cutter with trailer functionality.
Sending the harvester to a field with Giants or Autodrive helper.
The Trailer must be in a spot, where the combine has enough space to pick it up, reverse and then drive to the fieldwork course.
If there is not enough space, there is nothing we can do about it. Manuvering within an AI needs some space.
Unloading the header after finish work is planed but has *low priority*.
IMPORTANT: Right now there is a mod conflict with manual attach, this will be fixed in the manual attach mod in a later update.