fix: Better check for articulated axis vehicles
For some reason in newer FS releases all vehicles seem to have
and ArticulatedAxis, Crawlers and CrabSteering specialization,
regardless of the contents of the vehicle XML (that is,
regardless if the vehicle has that feature or not).

Add an extra check that should only return true if the
vehicle really has a real articulated axis.

#2656