fix: reduce fruit check area
When checking if there is currently fruit at a future
waypoint, use a smaller area (1x1 m instead of 5x3 m)
under the pipe to check for fruit.

Larger areas may include some of the fruit in the current
row and thus return a false positive even if the
row to the left is already harvested. This may also be
worsened by the resolution of the fruit map/angle of
row.

#2655