2+ Tippers 90% Reliable
Some edge cases to figure out. End of rows causes issues. Partially due to reliance on pipe in fruit map which isn't accurate for choppers. Need to create a better method for getting the chopper to send the tipper home as the current method resets after 3 secs which causes issues if the tipper is blocked on its way home