95% Feature Complete
Chase works. All features work. Some adjustments and tweaks need to be made. Need to still make a method to send the driver home instead of relying on the user input method(resets after 3 secs if there is some reason for the driver to go back into idle aka collision)