Attempts at Edge Case Fixes
We have to edge cases that aren't handle well. When the combine isn't aligned to the course proximity controller doesn't pick up the chopper. For now I have set the unloader to drive slowly in hopes that the chopper gets aligned and we find it again. Then the transition from connecting track to up/down. There is a moment when it isn't in either connecting state or turning state causing the pathfinder to go to this chopper position in between states which then it is to close to the Chopper to figure out what to do. I tried with isAboutToTurn() hopping that would bridge the gap but doesn't appear to work on connecting tracks