Attempt 5k on fixing edge cases
The backmarker node isn't correct on all choppers(Haven't tested others). Tried to fix it still can't get it to work. This breaks the code when we can't detect the Chopper as we fall back to localToLocal and since this point isn't correct we run into the chopper. Also the chopper doesn't stay in the Connecting Track state for the aligment turn to up/down which causes a miss calculation on the pathfinder due to wrong position received