AI Direction Fixes, PipeController Update
Replaced combineRootNode to AI Direction in drive besidesCombine as we are now measure pipeoffsetZ from the Direction node and not the root. This will cause misalignments. Removed the minus 2 as well as this was probably a fix to correct this issue. Move the calculation of the pipeOffset to its separate function. For choppers we want the pipeBaseNode as our reference as this point is fixed

Two known call stacks exist. End of chopper course doesn't exit cleanly. A combineToUnload is nil in idle can't find where we are calling this from an idle state