Some tweaks to changing unloader logic
Noticed a state where it couldn't decide whether to go or not. I am guessing it was the changing unloader logic as we now currently check to ensure we are only unloading to our current unloader in the unload state or a human/AD before driving.

Also another attempt at fixing the end of work call stack. It as to do with looking at waypoints beyond the end of the course. Made a new function in course to provide this check