Try to tweak those values a bit
Lookahead from 5.0 to 2.0 improoved it a lot and could be enough already. Same for resolution with 0.2
Tried go even finer with 1.0 and 0.1 wich works good as well, need to play arround with those values on more fields. Will only work well with new generator else some corners won't be generated as sharp corners.