Push Release 7.3.1.0
- fix for #2768
- fix for #2777
- fix for #2761 and other plow improvements.
- Improvement for reverse driving with offset tools (e.g. plows or potato harvester)
- Plows default noReverse configuration was removed, User who want them to not drive in reverse turn off "turn on field" or can also turn on "use pathfinder in turns".

- Added shovel mode back in. #2254

Please check the in-game help menu for how to set up the new mode.

We can't support all kinds of shovel vehicles (like bobcats) as they were built inconsistently (like the tool arms are sometimes flipped, and so on).

If it turns out to be a lot of non-working shovel vehicles, we'll consider implementing some kind of Tool Position Saver (like our separate mod)  or find some another solution, until then, we would prefer keeping the automatic control.

We were not able to test all different kinds of combinations. If you have a vehicle that does weird things, pls post a screenshot (with debug on) in our Issue: #2783