fix: reversing turn
When we have to reverse to reach the work start after the
turn, only change direction when we actually pass those
waypoints marked with CHANGE_DIRECTION_WHEN_ALIGNED instead
of checking the current waypoint.

This avoids switching the direction (and thus lowering the
implement) too early, when temporarily during the turn the
implement is aligned with the target row.