fix: Giants collision detection removed
We left the standard AIDriveStrategyCollision by
Giants in at the beginning as we did not have our
own solution. .

Now it seems to interfere with our ProximityController, stopping
the vehicle and not calling our strategy well before our
ProximityController would engage.

This practically disables all CP collision resolution
mechanisms, so better remove the Giants one and rely completely
on ours.