fix/refactor: bale finder pathfinder failure logic
Don't use the pathfinder controller retry logic as we don't
want to retry with the same goal.

Give up after 3 retries. Try another bale if the goal node
is invalid, or failed, but not close to the field edge.

If close to the field edge, back up a bit and start again.

Keep a list of bales tried to avoid retrying the same two bales
and causing a stack overflow.