Push Release 7.4.2.1
WARNING: While this version is backwards compatible with older savegames, it'll write the courses to the savegame in a new format. Vesions before 7.4.2.0 WILL NOT BE ABLE TO READ the games saved with this version! If you install this version, and decide to revert to an old one, YOU MUST BACK UP YOUR OLD SAVEGAME before!

- Even row width setting does work now #3423
- Fix for a strange waypoint transformation #3399
- Narrow field is now fixed too #3387
- Transition fix #3387
- Another fix for Bale Unloader #3381
- Fix for #3378
- Fix for plows not starting the work after headland #3413
- Vehicles in 180° turns near a diagonal field border are not leaving the field anymore.
- Fix for Multitool #3379

New for Multitool:
- A multitool now contains 2-5 single course for each lane depending on the amount of vehicles you selected.
- That course is switched when switching the lane.
- The old "symetric lane change" got renamed to "switching lanes" and hopefully is more clear about what it does and have to be selected in the generator before generating the course.
- Multitool courses saved by older versions will not work, you'll need to regenerate them!

Known Issue:
After reconnecting to the server, the selected lane setting for multitool gets reset and won't show the correct lane, but the tool keeps working on the correct one. This will be fixed in a later version, as we don't want to wait longer with important fixes.