fix: reverse at turn end
When a turn needs to be moved back to fit on the field,
it ends with a straight reversing section to back the
implement up to the work start point. The actual start
of this section depends on when the implement was aligned
with the row.

The generated start point of this reversing section however,
at this point, may lay well behind the implement's reversing
node, and this will result the PPC stuck in initializing mode
and the vehicle reversing forever, since it is moving further
and further away from the start point.

So, when changing to reverse, do not initialize to the first
reverse waypoint, instead, find the one closest to the
implement's reverser node.