fix: multitool headland turn
In Course:enrichWaypointData(), always use the non-offset position
of the waypoints. This fixes issues when rerunning enrichWaypointData()
on a course with offset.

Also, it wasn't correct in the first place, this is a chicken-egg problem,
since the offset calculation is based on the dx/dz values which at the
first run of enrichWaypointData() are not known as they are calculated
by enrichWaypointData().