More adjustments
No matter how much try arround with that, the most elegant solution would be having the distance between our reference waypoint and the pipe of the harvester/combine, be long enought to fit the unloader (in meter, as we don't know what unloader will unload the combine). With that known distance, we coould calculate how far we need to drive back, to then drive into the pocket and get our harvester straight.