fix: pipe measurement
Measure the pipe length after load and attach,
instead of when starting the driver to prevent
messing up vehicle states with the pipe measurement
hack.

Add two hidden vehicle settings for the pipe offset.

Seems to work for most, certainly not for the Oxbo BP2140e,
the x offset is way too low, may be a vehicle bug.