fix: Autodrive API
Removed the call to Autodrive's combineIsCallingDriver()
to check if the pipe needs to open. combineIsCallingDriver()
isn't doing anything other than checking
getIsCpHarvesterWaitingForUnload(), it isn't in any way
depends on the Autodrive state of the unloader, therefore,
it does not make a lot of sense, it just opens the pipe
unnecessarily.

Autodrive still can check if the CP harvester is waiting
for unload, and drive to the harvester. The trailer being
in the range will trigger opening the pipe.